requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=1.3.0
httpx>=0.24.0
openai>=1.0.0
groq>=0.4.0
gpt4all>=2.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
pollinations>=0.4.0
fpdf2>=2.7.0